requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.3
billiard==4.1.0
celery==5.3.4
certifi==2022.9.24
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.7
Django==4.1.7
django-admin-thumbnails==0.2.6
django-cors-headers==3.13.0
django-debug-toolbar==3.7.0
django-filter==22.1
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-writable-nested==0.7.0
drf-yasg==1.21.5
exceptiongroup==1.2.1
flower==2.0.1
humanize==4.9.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.2
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.2.0
pluggy==1.5.0
prometheus_client==0.20.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
pycparser==2.22
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.5
razorpay==1.3.0
redis==5.0.4
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tornado==6.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.8
whitenoise==6.5.0
